# use a separate go routine to write limiters