### Description
This is an alternative to #15642 
Use a go buffered go routine to accumulate Limiters in a transaction and write them to the immutable-radix-tree every 10Ms.

 

### Testing & Reproduction steps


#### Without any synchronization code in #15467 
``` bash
❯ go test -bench=BenchmarkTestRateLimiterAllowC -count 5 -benchtime=1s  -timeout 6000s -run=XXX ./agent/consul/multilimiter
goos: darwin
goarch: amd64
pkg: github.com/hashicorp/consul/agent/consul/multilimiter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
AC/no_prefill-16         	  556688	      2060 ns/op
AC/no_prefill-16         	  672924	      1964 ns/op
AC/no_prefill-16         	  627190	      1932 ns/op
AC/no_prefill-16         	  615247	      1952 ns/op
AC/no_prefill-16         	  661696	      1907 ns/op
AC/prefill_with_1K_keys-16         	  727677	      1894 ns/op
AC/prefill_with_1K_keys-16         	  611496	      2310 ns/op
AC/prefill_with_1K_keys-16         	  642594	      2043 ns/op
AC/prefill_with_1K_keys-16         	  590719	      1888 ns/op
AC/prefill_with_1K_keys-16         	  625422	      1937 ns/op
AC/prefill_with_10K_keys-16        	  494367	      2436 ns/op
AC/prefill_with_10K_keys-16        	  522981	      2491 ns/op
AC/prefill_with_10K_keys-16        	  529140	      2291 ns/op
AC/prefill_with_10K_keys-16        	  612369	      2356 ns/op
AC/prefill_with_10K_keys-16        	 1000000	      2607 ns/op
AC/prefill_with_100K_keys-16       	 1000000	      3285 ns/op
AC/prefill_with_100K_keys-16       	 1000000	      3626 ns/op
AC/prefill_with_100K_keys-16       	 1000000	      3891 ns/op
AC/prefill_with_100K_keys-16       	 1000000	      4479 ns/op
AC/prefill_with_100K_keys-16       	 1000000	      4352 ns/op
PASS
ok  	github.com/hashicorp/consul/agent/consul/multilimiter	66.699s
```

#### This branch (channel with transaction)

```bash
❯ go test -bench=BenchmarkTestRateLimiterAllowC -count 5 -benchtime=1s  -timeout 6000s -run=XXX ./agent/consul/multilimiter
goos: darwin
goarch: amd64
pkg: github.com/hashicorp/consul/agent/consul/multilimiter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
AC/no_prefill-16         	  465397	      3929 ns/op
AC/no_prefill-16         	  736290	      3168 ns/op
AC/no_prefill-16         	  770343	      2957 ns/op
AC/no_prefill-16         	  669290	      3759 ns/op
AC/no_prefill-16         	  823153	      3547 ns/op
AC/prefill_with_1K_keys-16         	  807282	      4839 ns/op
AC/prefill_with_1K_keys-16         	  718530	      2788 ns/op
AC/prefill_with_1K_keys-16         	  786652	      4330 ns/op
AC/prefill_with_1K_keys-16         	  465585	      2994 ns/op
AC/prefill_with_1K_keys-16         	  715992	      2619 ns/op
AC/prefill_with_10K_keys-16        	  699630	      2876 ns/op
AC/prefill_with_10K_keys-16        	  703555	      3843 ns/op
AC/prefill_with_10K_keys-16        	  347377	      3426 ns/op
AC/prefill_with_10K_keys-16        	  659928	      3561 ns/op
AC/prefill_with_10K_keys-16        	  575187	      2929 ns/op
AC/prefill_with_100K_keys-16       	  552110	      2893 ns/op
AC/prefill_with_100K_keys-16       	  538591	      3101 ns/op
AC/prefill_with_100K_keys-16       	  542137	      2986 ns/op
AC/prefill_with_100K_keys-16       	  501328	      3057 ns/op
AC/prefill_with_100K_keys-16       	  493551	      3292 ns/op
PASS
ok  	github.com/hashicorp/consul/agent/consul/multilimiter	176.243s
```