@dhiaayachi 2022-12-06
@boxofrad Thanks again for the review. I went through it and decided to use a sync.Once for the Run call. Please take another look when you have time.